+++ workers/unity/Packages/io.improbable.gdk.tools/.CodeGenTemplate/CodeGenerationLib/Model/Details/DetailsStore.cs
@@ -3,9 +3,12 @@ using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
+using System.IO;
 using System.Linq;
+using System.Text.RegularExpressions;
 using Improbable.Gdk.CodeGeneration.FileHandling;
 using Improbable.Gdk.CodeGeneration.Utils;
+using Newtonsoft.Json.Linq;
 using NLog;
 
 namespace Improbable.Gdk.CodeGeneration.Model.Details
@@ -16,6 +19,7 @@ namespace Improbable.Gdk.CodeGeneration.Model.Details
         public readonly IReadOnlyDictionary<string, UnityEnumDetails> Enums;
         public readonly IReadOnlyDictionary<string, UnityComponentDetails> Components;
         public readonly IReadOnlyList<ComponentSetDetails> ComponentSets;
+        public readonly IReadOnlyList<string> WorkerTypes;
 
         public readonly ImmutableHashSet<string> BlittableSet;
         public readonly IReadOnlyList<string> SchemaFiles;
@@ -27,7 +31,8 @@ namespace Improbable.Gdk.CodeGeneration.Model.Details
 
         private static readonly Logger Logger = LogManager.GetCurrentClassLogger();
 
-        public DetailsStore(SchemaBundle bundle, IEnumerable<string> serializationOverrides, IFileTree fileTree)
+        public DetailsStore(SchemaBundle bundle, IEnumerable<string> serializationOverrides, IFileTree fileTree,
+            string workerJsonDirectory)
         {
             this.bundle = bundle;
             FileTree = fileTree;
@@ -125,6 +130,8 @@ namespace Improbable.Gdk.CodeGeneration.Model.Details
             {
                 Logger.Trace($"Removed {numFieldsRemoved} recursive options.");
             }
+
+            WorkerTypes = ExtractWorkerTypes(workerJsonDirectory);
         }
 
         public HashSet<string> GetNestedTypes(string qualifiedName)
@@ -285,6 +292,36 @@ namespace Improbable.Gdk.CodeGeneration.Model.Details
             }
 
             return numFieldsRemoved;
+        }
+
+        private static List<string> ExtractWorkerTypes(string path)
+        {
+            Logger.Trace($"Extracting worker types from {path}.");
+
+            var workerTypes = new List<string>();
+
+            var fileNames = Directory.EnumerateFiles(path, "spatialos.*.worker.json").ToList();
+            Logger.Trace($"Found {fileNames.Count} worker json files:\n - {string.Join("\n - ", fileNames)}");
+
+            foreach (var fileName in fileNames)
+            {
+                Logger.Trace($"Extracting worker type from {fileName}.");
+                var match = Regex.Match(fileName, @"spatialos\.(.*)\.worker\.json");
+
+                if (!match.Success)
+                {
+                    Logger.Warn($"Unable to extract worker type from \"{fileName}\".");
+                    continue;
+                }
+
+                var workerType = match.Groups[1].Value;
+
+                Logger.Trace($"Adding {workerType} to list of worker types.");
+                workerTypes.Add(workerType);
+            }
+
+            Logger.Trace($"Found {workerTypes.Count} worker types:\n - {string.Join("\n - ", workerTypes)}");
+            return workerTypes;
         }
     }
 }